pyproject.toml
[project]
name = "batch-sales-transactions"
version = "0.1.0"
description = "Microservicio Django para procesamiento batch de transacciones de ventas"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "pytest-cov>=6.0.0",
    "factory-boy>=3.3.3",
]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=apps --cov-report=term-missing"
